move engine specific interfaces to n8n-core
