+++ packages/core/src/execution-engine/__tests__/active-workflows.test.ts
@@ -1,6 +1,5 @@
 import { mock } from 'jest-mock-extended';
 import type {
-	IGetExecuteTriggerFunctions,
 	INode,
 	ITriggerResponse,
 	IWorkflowExecuteAdditionalData,
@@ -18,6 +17,7 @@ import { ActiveWorkflows } from '../active-workflows';
 import type { PollContext } from '../node-execution-context';
 import type { ScheduledTaskManager } from '../scheduled-task-manager';
 import type { TriggersAndPollers } from '../triggers-and-pollers';
+import type { IGetExecuteTriggerFunctions } from '../interfaces';
 
 describe('ActiveWorkflows', () => {
 	const workflowId = 'test-workflow-id';
